# backend/shortVideoApiService/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /comment:
        post:
            tags:
                - CommentService
            description: 创建评论
            operationId: CommentService_CreateComment
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.CreateCommentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.CreateCommentResponse'
        delete:
            tags:
                - CommentService
            description: 删除评论
            operationId: CommentService_RemoveComment
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.RemoveCommentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.RemoveCommentResponse'
    /comment/child:
        post:
            tags:
                - CommentService
            operationId: CommentService_ListChildComment
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.ListChildCommentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.ListChildCommentResponse'
    /comment/video:
        post:
            tags:
                - CommentService
            description: 列出视频的评论
            operationId: CommentService_ListComment4Video
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.ListComment4VideoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.ListComment4VideoResponse'
...
    /follow:
        get:
            tags:
                - FollowService
            operationId: FollowService_ListFollowing
            parameters:
                - name: userId
                  in: query
                  description: '@gotags: json:"userId,omitempty,string"'
                  schema:
                    type: string
                - name: type
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: pagination.page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.size
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.ListFollowingResponse'
        post:
            tags:
                - FollowService
            operationId: FollowService_AddFollow
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.AddFollowRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.AddFollowResponse'
        delete:
            tags:
                - FollowService
            operationId: FollowService_RemoveFollow
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.RemoveFollowRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.RemoveFollowResponse'
...
components:
    schemas:
...
        svapi.AddFollowRequest:
            type: object
            properties:
                userId:
                    type: string
                    description: '@gotags: json:"userId,omitempty,string"'
        svapi.AddFollowResponse:
            type: object
            properties: {}
...
        svapi.Comment:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                videoId:
                    type: string
                    description: '@gotags: json:"videoId,omitempty,string"'
                parentId:
                    type: string
                    description: '@gotags: json:"parentId,omitempty,string"'
                user:
                    $ref: '#/components/schemas/svapi.CommentUser'
                replyUser:
                    $ref: '#/components/schemas/svapi.CommentUser'
                content:
                    type: string
                date:
                    type: string
                likeCount:
                    type: string
                replyCount:
                    type: string
                comments:
                    type: array
                    items:
                        $ref: '#/components/schemas/svapi.Comment'
        svapi.CommentUser:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                name:
                    type: string
                avatar:
                    type: string
                isFollowing:
                    type: boolean
...
        svapi.CreateCommentRequest:
            type: object
            properties:
                videoId:
                    type: string
                    description: '@gotags: json:"videoId,omitempty,string"'
                content:
                    type: string
                parentId:
                    type: string
                    description: '@gotags: json:"parentId,omitempty,string"'
                replyUserId:
                    type: string
                    description: '@gotags: json:"replyUserId,omitempty,string"'
        svapi.CreateCommentResponse:
            type: object
            properties:
                comment:
                    $ref: '#/components/schemas/svapi.Comment'
...
        svapi.FollowUser:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                name:
                    type: string
                avatar:
                    type: string
                isFollowing:
                    type: boolean
...
        svapi.ListChildCommentRequest:
            type: object
            properties:
                commentId:
                    type: string
                    description: '@gotags: json:"commentId,omitempty,string"'
                pagination:
                    $ref: '#/components/schemas/svapi.PaginationRequest'
        svapi.ListChildCommentResponse:
            type: object
            properties:
                comments:
                    type: array
                    items:
                        $ref: '#/components/schemas/svapi.Comment'
                pagination:
                    $ref: '#/components/schemas/svapi.PaginationResponse'
...
        svapi.ListComment4VideoRequest:
            type: object
            properties:
                videoId:
                    type: string
                    description: '@gotags: json:"videoId,omitempty,string"'
                pagination:
                    $ref: '#/components/schemas/svapi.PaginationRequest'
        svapi.ListComment4VideoResponse:
            type: object
            properties:
                comments:
                    type: array
                    items:
                        $ref: '#/components/schemas/svapi.Comment'
                pagination:
                    $ref: '#/components/schemas/svapi.PaginationResponse'
...
        svapi.ListFollowingResponse:
            type: object
            properties:
                users:
                    type: array
                    items:
                        $ref: '#/components/schemas/svapi.FollowUser'
                pagination:
                    $ref: '#/components/schemas/svapi.PaginationResponse'
...
        svapi.PaginationRequest:
            type: object
            properties:
                page:
                    type: integer
                    format: int32
                size:
                    type: integer
                    format: int32
                sort:
                    type: array
                    items:
                        $ref: '#/components/schemas/svapi.SortField'
        svapi.PaginationResponse:
            type: object
            properties:
                page:
                    type: integer
                    format: int32
                total:
                    type: integer
                    format: int32
                count:
                    type: integer
                    format: int32
...
        svapi.RemoveCommentRequest:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
        svapi.RemoveCommentResponse:
            type: object
            properties: {}
...
        svapi.RemoveFollowRequest:
            type: object
            properties:
                userId:
                    type: string
                    description: '@gotags: json:"userId,omitempty,string"'
        svapi.RemoveFollowResponse:
            type: object
            properties: {}
...
        svapi.SortField:
            type: object
            properties:
                field:
                    type: string
                order:
                    type: integer
                    format: enum
...